+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 (not yet released)
 
+### Security fixes
+
+* Fixed a vunlerability in the `CHANGESET_REQ` message handler that allowed a
+  user with any access to read any pad if the pad ID is known.
+
 ### Notable enhancements and fixes
 
 * Fixed a bug that caused all pad edit messages received at the server to go